move globals to blitz core, move event listener to blitz-next
