+++ packages/blitz-next/src/global.ts
@@ -1,10 +1,5 @@
 import {QueryClient} from "@tanstack/react-query"
-import {BeforeHttpRequest, BeforeHttpResponse, OnRpcErrorHook, OnSessionCreatedHook} from "blitz"
 
 declare global {
   var queryClient: QueryClient
-  var __BLITZ_beforeHttpRequest: BeforeHttpRequest
-  var __BLITZ_beforeHttpResponse: BeforeHttpResponse
-  var __BLITZ_onRpcError: OnRpcErrorHook
-  var __BLITZ_onSessionCreated: OnSessionCreatedHook
 }
